Fixes for re-enabling more MSVC level 4 warnings: base/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch
* Assignment inside conditional
* Nonstandard extension: unnamed struct/union
* Octal escape sequence terminated by decimal number
* Possibly-uninitialized local variable

This also contains a very small number of other cleanups/simplifications to
nearby code.

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/368133002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281204 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ed8e57da499e4ad0398f78696c3b2dfaa66ceaba

+++ base/i18n/break_iterator.cc
@@ -12,7 +12,7 @@
 namespace base {
 namespace i18n {
 
-const size_t npos = -1;
+const size_t npos = static_cast<size_t>(-1);
 
 BreakIterator::BreakIterator(const string16& str, BreakType break_type)
     : iter_(NULL),
